# api-contract/openapi.yaml
openapi: 3.0.3
info:
    title: SafeGuard API
    description: SafeGuard 微信小程序后端API服务
    version: 1.0.0
    contact:
        name: SafeGuard Development Team
        email: dev@safeguard.com
servers:
    - url: http://localhost:9999/api
      description: 开发环境
    - url: https://prod.safeguard.com/api
      description: 生产环境

# 通用组件定义
components:
    schemas:
...
        StandardResponse:
            type: object
            properties:
                code:
                    type: integer
                    description: 1-成功，0-失败
                    example: 1
                msg:
                    type: string
                    description: 状态消息
                    example: "success"
                data:
                    type: object
                    description: 响应数据
                    nullable: true

        # 错误响应
...
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT

# 全局安全要求
security:
    - BearerAuth: []

# API路径定义
paths:
...
    /api/checkin/create:
        post:
            tags:
                - 分享模块
            summary: 创建可分享的打卡邀请链接
            description: 为打卡记录创建可分享的链接
            security:
                - BearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - rule_id
                            properties:
                                rule_id:
                                    type: integer
                                    description: 打卡规则ID
                                expire_hours:
                                    type: integer
                                    default: 168
                                    description: 链接过期时间（小时）
            responses:
                "200":
                    description: 创建成功
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: "#/components/schemas/StandardResponse"
                                    - type: object
                                      properties:
                                          data:
                                              type: object
                                              properties:
                                                  token:
                                                      type: string
                                                  url:
                                                      type: string
                                                  mini_path:
                                                      type: string
                                                  expire_at:
                                                      type: string
                                                      format: date-time
    /api/checkin/resolve:
        get:
            tags:
                - 分享模块
            summary: 解析分享链接
            description: 解析分享链接，返回打卡规则信息
            security: []
            parameters:
                - name: token
                  in: query
                  required: true
                  schema:
                      type: string
                  description: 分享token
            responses:
                "200":
                    description: 解析成功
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: "#/components/schemas/StandardResponse"
                                    - type: object
                                      properties:
                                          data:
                                              type: object
                                              properties:
                                                  rule_info:
                                                      type: object
                                                      properties:
                                                          rule_id:
                                                              type: integer
                                                          title:
                                                              type: string
                                                          description:
                                                              type: string
                                                          checkin_time:
                                                              type: string
                                                          repeat_days:
                                                              type: array
                                                              items:
                                                                  type: integer
                                                          is_enabled:
                                                              type: boolean
                                                  share_info:
                                                      type: object
                                                      properties:
                                                          share_user_id:
                                                              type: integer
                                                          share_user_nickname:
                                                              type: string
                                                          share_user_avatar:
                                                              type: string
                                                          created_at:
                                                              type: string
                                                              format: date-time
                                                          expires_at:
                                                              type: string
                                                              format: date-time
                                                          access_count:
                                                              type: integer
...